% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_controls.R
\name{check_controls}
\alias{check_controls}
\title{Check \code{controls}.}
\usage{
check_controls(controls)
}
\arguments{
\item{controls}{Either \code{NULL} or a named list with the following elements. Missing
elements are set to default values (printed in parentheses).
\itemize{
\item \code{init_runs} \code{(5)}:
The number of short initial searches.
\item \code{init_min} \code{(-1)}:
The minimum value for the random initialization.
\item \code{init_max} \code{(1)}:
The maximum value for the random initialization.
\item \code{init_iterlim} \code{(20)}:
The number of iterations for the short initial searches.
\item \code{neighborhoods} \code{(5)}:
The number of nested neighborhoods.
\item \code{neighbors} \code{(5)}:
The number of neighbors in each neighborhood.
\item \code{beta} \code{(0.05)}:
A non-negative weight factor to account for the function's curvature in the
selection of the neighbors. If \code{beta = 0}, the curvature is ignored.
The higher the value, the higher probability of selecting a neighbor in the
direction of the highest function curvature.
\item \code{iterlim} \code{(1000)}:
The maximum number of iterations to be performed before the local search is
terminated.
\item \code{only_global} \code{(FALSE)}:
A boolean. Set \code{only_global = TRUE} if you are only interested in the
global optimum. In this case, the algorithm will interrupt any local search
prematurely that converges to a local optimum. This saves computation time.
If \code{only_global = FALSE} (the default), the algorithm also looks for local
optima.
\item \code{tolerance} \code{(1e-5)}:
Passed on to \code{\link{all.equal}}, which is used to check if two optima
are equal.
}}
}
\value{
The checked and filled list \code{controls}.
}
\description{
Function that checks \code{controls} for \code{\link{search}}.
}
