% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_f.R
\name{check_f}
\alias{check_f}
\title{Check \code{f}.}
\usage{
check_f(f, npar, controls)
}
\arguments{
\item{f}{A function that computes value, gradient, and Hessian of the function to be
optimized and returns them as a named list with elements \code{value},
\code{gradient}, and \code{hessian}. It is required that \code{hessian} is
of class \code{matrix}.}

\item{npar}{The number of parameters of \code{f}.}
}
\value{
Invisibly a boolean, if \code{TRUE} than \code{f} is proper, \code{FALSE}
otherwise.
}
\description{
Function that checks \code{f} for \code{\link{search}}.
}
