% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{local}
\alias{local}
\title{Perform trust region local search.}
\usage{
local(f, parinit, minimize, controls, L)
}
\arguments{
\item{f}{A function that computes value, gradient, and Hessian of the function to be
optimized and returns them as a named list with elements \code{value},
\code{gradient}, and \code{hessian}. It is required that \code{hessian} is
of class \code{matrix}.}

\item{parinit}{Passed on to \code{\link[trust]{trust}}.}

\item{L}{A list of identified optima which contains lists with
\itemize{
\item \code{value} and
\item \code{argument}
}
of each identified optimum.}
}
\value{
A list of
\itemize{
\item \code{success}: A boolean, determining wether the local search
successfully converged.
\item \code{value}: The value at the point where the local search
terminated.
\item \code{argument}: The point where the local search terminated.
}
}
\description{
Function that links to \code{\link[trust]{trust}}.
}
