% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_output.R
\name{prepare_output}
\alias{prepare_output}
\title{Prepare output list.}
\usage{
prepare_output(L, minimize)
}
\arguments{
\item{L}{A list of identified optima which contains lists with
\itemize{
\item \code{value} and
\item \code{argument}
}
of each identified optimum.}

\item{minimize}{A boolean, determining whether \code{f} should be minimized
\code{minimize = TRUE} (the default) or maximized \code{minimize = FALSE}.}
}
\value{
A list of two list,
\itemize{
\item \code{global} optima and
\item \code{local} optima,
}
each of which contains lists with
\itemize{
\item \code{value} and
\item \code{argument}
}
of each identified optimum.
}
\description{
A function that prepares the output of \code{\link{search}} by distinguishing
between local and global optima in \code{L}.
}
