% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.locglob.R
\name{print.locglob}
\alias{print.locglob}
\title{Print method for \code{locglob}.}
\usage{
\method{print}{locglob}(x, ...)
}
\arguments{
\item{x}{The output of \code{\link{search}}, which is an object of class
\code{locglob}.}

\item{...}{ignored}
}
\description{
Print method for \code{locglob}.
}
