% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{unique}
\alias{unique}
\title{Check new optimum for uniqueness}
\usage{
unique(L, argument, tolerance)
}
\arguments{
\item{L}{A list of identified optima which contains lists with
\itemize{
\item \code{value} and
\item \code{argument}
}
of each identified optimum.}

\item{argument}{The argument of a candidate optimum.}

\item{tolerance}{A non-negative numeric value. For an identified optimum and a candidate
optimum, if all coordinate differences are smaller than \code{tolerance}, they
are considered as equal.
\code{tolerance},}
}
\value{
A boolean. If \code{TRUE}, \code{argument} is not contained in \code{L}.
If \code{FALSE}, \code{argument} is already contained in \code{L}.
}
\description{
Function that checks if a new optimum at \code{argument} is not yet contained
in \code{L}.
}
