% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vntrs-package.R
\docType{package}
\name{vntrs-package}
\alias{vntrs-package}
\alias{_PACKAGE}
\title{vntrs: Variable Neighborhood Trust Region Search}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An implementation of the variable neighborhood trust region algorithm Bierlaire et al. (2009) "A Heuristic for Nonlinear Global Optimization" \doi{10.1287/ijoc.1090.0343}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://loelschlaeger.de/vntrs/}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{lennart.oelschlaeger@uni-bielefeld.de} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

}
\keyword{internal}
