% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vntrs.R
\name{vntrs}
\alias{vntrs}
\title{Perform variable neighborhood trust region search.}
\usage{
vntrs(target, npar, vntrs_controls)
}
\arguments{
\item{target}{A function that computes value, gradient, and Hessian of the function to be optimized and returns them as a list with components \code{value}, \code{gradient}, and \code{hessian}.}

\item{npar}{The number of parameters of \code{target}.}

\item{vntrs_controls}{A list of controls.}
}
\value{
The point at which the optimal value of \code{target} is obtained.
}
\description{
Function that performs variable neighborhood trust region search.
}
\details{
See the README file for details and examples on how to specify \code{vntrs_controls}.
}
\references{
This is an implementation of the heuristic presented in "A Heuristic for Nonlinear Global Optimization" (Bierlaire et al., 2009) \url{https://doi.org/10.1287/ijoc.1090.0343}.
}
