% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vntrs_check_controls.R
\name{vntrs_check_controls}
\alias{vntrs_check_controls}
\title{Check controls for the variable neighborhood trust region search.}
\usage{
vntrs_check_controls(vntrs_controls)
}
\arguments{
\item{vntrs_controls}{A list of controls.}
}
\value{
\code{vntrs_controls}
}
\description{
Function that checks the controls for the variable neighborhood trust region search.
}
\details{
See the README file for details.
}
