% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vntrs_initialize.R
\name{vntrs_initialize}
\alias{vntrs_initialize}
\title{Initialize variable neighborhood trust region search.}
\usage{
vntrs_initialize(target, npar, vntrs_controls)
}
\arguments{
\item{target}{A function that computes value, gradient, and Hessian of the function to be optimized and returns them as a list with components \code{value}, \code{gradient}, and \code{hessian}.}

\item{npar}{The number of parameters of \code{target}.}

\item{vntrs_controls}{A list of controls.}
}
\value{
A list of the set of identified local optima and best initial parameter values.
}
\description{
Function that initializes the variable neighborhood trust region search.
}
