% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vntrs_local_search.R
\name{vntrs_local_search}
\alias{vntrs_local_search}
\title{Perform local search based on a trust-region framework.}
\usage{
vntrs_local_search(target, par_init, iterlim, interrupt_rate, minimize, L)
}
\arguments{
\item{target}{A function that computes value, gradient, and Hessian of the function to be optimized and returns them as a list with components \code{value}, \code{gradient}, and \code{hessian}.}

\item{iterlim}{A positive integer specifying the maximum number of iterations to be performed before the local search is terminated.}

\item{interrupt_rate}{A numeric between 0 and 1, determining the rate to check for premature interruption.}

\item{L}{}
}
\value{

}
\description{
Function that performs local search based on a trust-region framework.
}
\details{
The trust-region framework is provided by \link[trust]{trust}.
}
