% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vntrs_neighborhood.R
\name{vntrs_neighborhood}
\alias{vntrs_neighborhood}
\title{Select neighborhood.}
\usage{
vntrs_neighborhood(target, x, neighborhood_size, vntrs_controls)
}
\arguments{
\item{target}{A function that computes value, gradient, and Hessian of the function to be optimized and returns them as a list with components \code{value}, \code{gradient}, and \code{hessian}.}

\item{x}{A point in the domain of \code{target}.}

\item{neighborhood_size}{A scaling factor specifying the size of the neighborhood.}

\item{vntrs_controls}{A list of controls for the variable neighborhood trust region search.}
}
\value{
A list of neighbors.
}
\description{
Function that selects neighbors around a given point.
}
