% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interruption.R
\name{interruption}
\alias{interruption}
\title{Interrupt local search.}
\usage{
interruption(f, point, L, minimize)
}
\arguments{
\item{f}{A function that computes value, gradient, and Hessian of the function to be
optimized and returns them as a named list with elements \code{value},
\code{gradient}, and \code{hessian}.}

\item{point}{The current location of the local search.}

\item{L}{A list of identified optima which contains lists with
\itemize{
\item \code{value} and
\item \code{argument}
}
of each identified optimum.}

\item{minimize}{If \code{TRUE}, \code{f} gets minimized. If \code{FALSE}, maximized.}
}
\value{
\code{TRUE} for premature interruption, \code{FALSE} if not.
}
\description{
This function checks if the local search can be interrupted prematurely.
}
