% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.vntrs.R
\name{print.vntrs}
\alias{print.vntrs}
\title{Print method for an object of class \code{vntrs}.}
\usage{
\method{print}{vntrs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{vntrs}.}

\item{...}{Ignored.}
}
\value{
Returns invisibly \code{x}.
}
\description{
Print method for an object of class \code{vntrs}.
}
